fix: cast base_price in service catalog seed

+++ supabase/migrations/090_seed_service_catalog.sql
@@ -25,7 +25,7 @@ SELECT
   bill.id AS billing_type_id,
   unit.id AS unit_id,
   vat.id AS vat_rate_id,
-  v.base_price,
+  v.base_price::numeric,
   v.active,
   v.note
 FROM (
